+++ backend/src/main/java/net/pengcook/recipe/controller/RecipeController.java
@@ -14,6 +14,7 @@ import net.pengcook.recipe.dto.RecipeResponse;
 import net.pengcook.recipe.dto.RecipeStepRequest;
 import net.pengcook.recipe.dto.RecipeStepResponse;
 import net.pengcook.recipe.service.RecipeService;
+import net.pengcook.recipe.service.RecipeStepService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -30,6 +31,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class RecipeController {
 
     private final RecipeService recipeService;
+    private final RecipeStepService recipeStepService;
 
     @GetMapping
     public List<MainRecipeResponse> readRecipes(@ModelAttribute @Valid PageRecipeRequest pageRecipeRequest) {
@@ -40,25 +42,6 @@ public class RecipeController {
     @ResponseStatus(HttpStatus.CREATED)
     public RecipeResponse createRecipe(@LoginUser UserInfo userInfo, @RequestBody @Valid RecipeRequest recipeRequest) {
         return recipeService.createRecipe(userInfo, recipeRequest);
-    }
-
-    @GetMapping("/{recipeId}/steps")
-    public List<RecipeStepResponse> readRecipeSteps(@PathVariable long recipeId) {
-        return recipeService.readRecipeSteps(recipeId);
-    }
-
-    @GetMapping("/{recipeId}/steps/{sequence}")
-    public RecipeStepResponse readRecipeStep(@PathVariable long recipeId, @PathVariable long sequence) {
-        return recipeService.readRecipeStep(recipeId, sequence);
-    }
-
-    @PostMapping("/{recipeId}/steps")
-    @ResponseStatus(HttpStatus.CREATED)
-    public RecipeStepResponse createRecipeStep(
-            @PathVariable long recipeId,
-            @RequestBody @Valid RecipeStepRequest recipeStepRequest
-    ) {
-        return recipeService.createRecipeStep(recipeId, recipeStepRequest);
     }
 
     @GetMapping("/search")
@@ -75,4 +58,22 @@ public class RecipeController {
         return recipeService.readRecipesOfUser(recipeOfUserRequest);
     }
 
+    @GetMapping("/{recipeId}/steps")
+    public List<RecipeStepResponse> readRecipeSteps(@PathVariable long recipeId) {
+        return recipeStepService.readRecipeSteps(recipeId);
+    }
+
+    @GetMapping("/{recipeId}/steps/{sequence}")
+    public RecipeStepResponse readRecipeStep(@PathVariable long recipeId, @PathVariable long sequence) {
+        return recipeStepService.readRecipeStep(recipeId, sequence);
+    }
+
+    @PostMapping("/{recipeId}/steps")
+    @ResponseStatus(HttpStatus.CREATED)
+    public RecipeStepResponse createRecipeStep(
+            @PathVariable long recipeId,
+            @RequestBody @Valid RecipeStepRequest recipeStepRequest
+    ) {
+        return recipeStepService.createRecipeStep(recipeId, recipeStepRequest);
+    }
 }
