+++ backend/src/main/java/net/pengcook/recipe/service/RecipeService.java
@@ -1,11 +1,9 @@
 package net.pengcook.recipe.service;
 
 import java.time.LocalTime;
-import java.time.format.DateTimeParseException;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import net.pengcook.authentication.domain.UserInfo;
@@ -15,7 +13,6 @@ import net.pengcook.category.service.CategoryService;
 import net.pengcook.image.service.S3ClientService;
 import net.pengcook.ingredient.service.IngredientService;
 import net.pengcook.recipe.domain.Recipe;
-import net.pengcook.recipe.domain.RecipeStep;
 import net.pengcook.recipe.dto.AuthorResponse;
 import net.pengcook.recipe.dto.CategoryResponse;
 import net.pengcook.recipe.dto.IngredientResponse;
@@ -25,12 +22,8 @@ import net.pengcook.recipe.dto.RecipeDataResponse;
 import net.pengcook.recipe.dto.RecipeOfUserRequest;
 import net.pengcook.recipe.dto.RecipeRequest;
 import net.pengcook.recipe.dto.RecipeResponse;
-import net.pengcook.recipe.dto.RecipeStepRequest;
-import net.pengcook.recipe.dto.RecipeStepResponse;
 import net.pengcook.recipe.exception.InvalidParameterException;
-import net.pengcook.recipe.exception.NotFoundException;
 import net.pengcook.recipe.repository.RecipeRepository;
-import net.pengcook.recipe.repository.RecipeStepRepository;
 import net.pengcook.user.domain.User;
 import net.pengcook.user.repository.UserRepository;
 import org.springframework.data.domain.PageRequest;
@@ -44,7 +37,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class RecipeService {
 
     private final RecipeRepository recipeRepository;
-    private final RecipeStepRepository recipeStepRepository;
     private final CategoryRecipeRepository categoryRecipeRepository;
     private final UserRepository userRepository;
 
@@ -53,7 +45,7 @@ public class RecipeService {
     private final S3ClientService s3ClientService;
 
     public List<MainRecipeResponse> readRecipes(PageRecipeRequest pageRecipeRequest) {
-        Pageable pageable = getPageable(pageRecipeRequest.pageNumber(), pageRecipeRequest.pageSize());
+        Pageable pageable = getValidatedPageable(pageRecipeRequest.pageNumber(), pageRecipeRequest.pageSize());
         List<Long> recipeIds = recipeRepository.findRecipeIds(pageable);
 
         List<RecipeDataResponse> recipeDataResponses = recipeRepository.findRecipeData(recipeIds);
@@ -79,40 +71,9 @@ public class RecipeService {
         return new RecipeResponse(savedRecipe);
     }
 
-    public List<RecipeStepResponse> readRecipeSteps(long recipeId) {
-        List<RecipeStep> recipeSteps = recipeStepRepository.findAllByRecipeIdOrderBySequence(recipeId);
-        return convertToRecipeStepResponses(recipeSteps);
-    }
-
-    public RecipeStepResponse readRecipeStep(long recipeId, long sequence) {
-        RecipeStep recipeStep = recipeStepRepository.findByRecipeIdAndSequence(recipeId, sequence)
-                .orElseThrow(() -> new NotFoundException("해당되는 레시피 스텝 정보가 없습니다."));
-
-        return new RecipeStepResponse(recipeStep);
-    }
-
-    public RecipeStepResponse createRecipeStep(long recipeId, RecipeStepRequest recipeStepRequest) {
-        validateRecipeStepSequence(recipeId, recipeStepRequest.sequence());
-        Recipe recipe = getRecipeByRecipeId(recipeId);
-        String imageUrl = getValidatedImageUrl(recipeStepRequest.image());
-        String description = recipeStepRequest.description();
-        LocalTime cookingTime = getValidatedCookingTime(recipeStepRequest.cookingTime());
-
-        Optional<RecipeStep> existingRecipeStep = recipeStepRepository.findByRecipeIdAndSequence(
-                recipeId,
-                recipeStepRequest.sequence()
-        );
-
-        RecipeStep recipeStep = existingRecipeStep
-                .map(currentRecipeStep -> currentRecipeStep.update(imageUrl, description, cookingTime))
-                .orElseGet(() -> saveRecipeStep(recipe, imageUrl, recipeStepRequest, cookingTime));
-
-        return new RecipeStepResponse(recipeStep);
-    }
-
     public List<MainRecipeResponse> readRecipesOfCategory(RecipeOfCategoryRequest request) {
         String categoryName = request.category();
-        Pageable pageable = getPageable(request.pageNumber(), request.pageSize());
+        Pageable pageable = getValidatedPageable(request.pageNumber(), request.pageSize());
         List<Long> recipeIds = categoryRecipeRepository.findRecipeIdsByCategoryName(categoryName, pageable);
 
         List<RecipeDataResponse> recipeDataResponses = recipeRepository.findRecipeData(recipeIds);
@@ -121,17 +82,11 @@ public class RecipeService {
 
     public List<MainRecipeResponse> readRecipesOfUser(RecipeOfUserRequest request) {
         long userId = request.userId();
-        Pageable pageable = getPageable(request.pageNumber(), request.pageSize());
+        Pageable pageable = getValidatedPageable(request.pageNumber(), request.pageSize());
         List<Long> recipeIds = recipeRepository.findRecipeIdsByUserId(userId, pageable);
 
         List<RecipeDataResponse> recipeDataResponses = recipeRepository.findRecipeData(recipeIds);
         return convertToMainRecipeResponses(recipeDataResponses);
-    }
-
-    private List<RecipeStepResponse> convertToRecipeStepResponses(List<RecipeStep> recipeSteps) {
-        return recipeSteps.stream()
-                .map(RecipeStepResponse::new)
-                .toList();
     }
 
     private List<MainRecipeResponse> convertToMainRecipeResponses(List<RecipeDataResponse> recipeDataResponses) {
@@ -178,60 +133,11 @@ public class RecipeService {
                 .collect(Collectors.toList());
     }
 
-    private Pageable getPageable(int pageNumber, int pageSize) {
+    private Pageable getValidatedPageable(int pageNumber, int pageSize) {
         long offset = (long) pageNumber * pageSize;
         if (offset > Integer.MAX_VALUE) {
             throw new InvalidParameterException("적절하지 않은 페이지 정보입니다.");
         }
         return PageRequest.of(pageNumber, pageSize);
-    }
-
-    private Recipe getRecipeByRecipeId(long recipeId) {
-        return recipeRepository.findById(recipeId)
-                .orElseThrow(() -> new NotFoundException("해당되는 레시피가 없습니다."));
-    }
-
-    private void validateRecipeStepSequence(long recipeId, int sequence) {
-        int previousSequence = sequence - 1;
-        if (previousSequence >= 1) {
-            recipeStepRepository.findByRecipeIdAndSequence(recipeId, previousSequence)
-                    .orElseThrow(() -> new InvalidParameterException("이전 스텝이 등록되지 않았습니다."));
-        }
-    }
-
-    private RecipeStep saveRecipeStep(
-            Recipe recipe, String imageUrl,
-            RecipeStepRequest recipeStepRequest,
-            LocalTime cookingTime
-    ) {
-        RecipeStep recipeStep = new RecipeStep(
-                recipe,
-                imageUrl,
-                recipeStepRequest.description(),
-                recipeStepRequest.sequence(),
-                cookingTime
-        );
-
-        return recipeStepRepository.save(recipeStep);
-    }
-
-    private String getValidatedImageUrl(String image) {
-        if (image == null) {
-            return null;
-        }
-        if (image.isBlank()) {
-            throw new InvalidParameterException("적절하지 않은 이미지 이름입니다.");
-        }
-        return s3ClientService.getImageUrl(image).url();
-    }
-
-    private LocalTime getValidatedCookingTime(String cookingTime) {
-        try {
-            return Optional.ofNullable(cookingTime)
-                    .map(LocalTime::parse)
-                    .orElse(null);
-        } catch (DateTimeParseException exception) {
-            throw new InvalidParameterException("적절하지 않은 조리시간입니다.");
-        }
     }
 }
